Persist webhooks and move redis store into its own file
